Edge-Receive-Notify/requirements.txt
azure-iot-device==2.12.0
azure-iot-hub==2.6.1
wget==3.2